+++ packages/sdk-js/src/adapter-sveltekit/ast-transforms/+layout.js.ts
@@ -1,11 +1,14 @@
 import { dedent } from "ts-dedent"
 import type { SourceFile } from "ts-morph"
-import { assertNoImportsFromSdkJs } from "../../ast-transforms/assertions.js"
-import { addImport, isOptOutImportPresent } from "../../ast-transforms/utils/imports.js"
+import {
+	addImport,
+	isOptOutImportPresent,
+	isSdkImportPresent,
+	removeImport,
+} from "../../ast-transforms/utils/imports.js"
 import { wrapExportedFunction } from "../../ast-transforms/utils/wrap.js"
 import { codeToSourceFile, nodeToCode } from "../../ast-transforms/utils/js.util.js"
 import type { TransformConfig } from "../vite-plugin/config.js"
-import { filePathForOutput } from "../vite-plugin/fileInformation.js"
 
 // ------------------------------------------------------------------------------------------------
 
@@ -17,7 +20,7 @@ const addImports = (
 	wrapperFunctionName: string,
 ) => {
 	addImport(sourceFile, "@inlang/sdk-js/adapter-sveltekit/shared", wrapperFunctionName)
-	if (!config.languageInUrl) {
+	if (root && !config.languageInUrl) {
 		addImport(sourceFile, "$app/environment", "browser")
 		addImport(
 			sourceFile,
@@ -60,16 +63,15 @@ export const transformLayoutJs = (
 	const sourceFile = codeToSourceFile(code, filePath)
 
 	if (isOptOutImportPresent(sourceFile)) return code
+	if (!root && !isSdkImportPresent(sourceFile)) return code
 
-	assertNoImportsFromSdkJs(sourceFile, filePathForOutput(config, filePath), "+layout.js", root) // TODO: implement functionality
-	if (!root) return code // for now we don't need to transform non-root files
-
-	const wrapperFunctionName = root ? "initRootLayoutLoadWrapper" : "initLayoutLoadWrapper"
+	const wrapperFunctionName = root ? "initRootLayoutLoadWrapper" : "initLoadWrapper"
 
 	addImports(sourceFile, config, root, wrapperFunctionName)
 
 	const options = root ? getOptions(config, root) : undefined
 	wrapExportedFunction(sourceFile, options, wrapperFunctionName, "load")
+	removeImport(sourceFile, "@inlang/sdk-js")
 
 	return nodeToCode(sourceFile)
 }