+++ packages/sdk-js/src/adapter-sveltekit/ast-transforms/+page.server.js.ts
@@ -1,28 +1,25 @@
-import { assertNoImportsFromSdkJs } from "../../ast-transforms/assertions.js"
-import { isOptOutImportPresent } from "../../ast-transforms/utils/imports.js"
-import { codeToSourceFile } from "../../ast-transforms/utils/js.util.js"
+import {
+	addImport,
+	isOptOutImportPresent,
+	isSdkImportPresent,
+	removeImport,
+} from "../../ast-transforms/utils/imports.js"
+import { codeToSourceFile, nodeToCode } from "../../ast-transforms/utils/js.util.js"
+import { wrapExportedFunction } from "../../ast-transforms/utils/wrap.js"
 import type { TransformConfig } from "../vite-plugin/config.js"
-import { filePathForOutput } from "../vite-plugin/fileInformation.js"
 
-export const transformPageServerJs = (
-	filePath: string,
-	config: TransformConfig,
-	code: string,
-	root: boolean,
-) => {
+export const transformPageServerJs = (filePath: string, config: TransformConfig, code: string) => {
 	const sourceFile = codeToSourceFile(code, filePath)
 
 	if (isOptOutImportPresent(sourceFile)) return code
+	if (!isSdkImportPresent(sourceFile)) return code
 
-	assertNoImportsFromSdkJs(sourceFile, filePathForOutput(config, filePath), "+page.server.js", root) // TODO: implement functionality
+	const wrapperFunctionName = "initServerLoadWrapper"
 
-	return code // for now we don't need to transform any files
+	addImport(sourceFile, "@inlang/sdk-js/adapter-sveltekit/server", wrapperFunctionName)
 
-	// const wrapperFunctionName = 'initServerLoadWrapper'
+	wrapExportedFunction(sourceFile, undefined, wrapperFunctionName, "load")
+	removeImport(sourceFile, "@inlang/sdk-js")
 
-	// addImport(sourceFile, '@inlang/sdk-js/adapter-sveltekit/server', wrapperFunctionName)
-
-	// wrapExportedFunction(sourceFile, undefined, wrapperFunctionName, 'load')
-
-	// return nodeToCode(sourceFile)
+	return nodeToCode(sourceFile)
 }